+++ src/app/components/new-customer/new-customer.component.ts
@@ -1,13 +1,20 @@
 import { Component, OnInit } from '@angular/core';
 import { NgForm } from '@angular/forms';
+import { Router } from '@angular/router';
+import { CustomersService } from '../../services/customers.service';
 import { Customer } from '../../interfaces/customer';
+
 @Component({
   selector: 'app-new-customer',
   templateUrl: './new-customer.component.html',
   styleUrls: ['./new-customer.component.scss'],
 })
 export class NewCustomerComponent implements OnInit {
-  constructor() {}
+  constructor(
+    private routerService: Router,
+    private CustomersService: CustomersService
+  ) {}
+
   form = {
     firstName: '',
     lastName: '',
@@ -28,9 +35,11 @@ export class NewCustomerComponent implements OnInit {
     });
   }
 
-  onSubmit({ value, valid }: { value: Customer; valid: boolean }): void {
-    console.log(value, valid);
-    // TODO: insert document to db
+  async onSubmit({ value, valid }: { value: Customer; valid: boolean }) {
+    if (valid) {
+      await this.CustomersService.add(value);
+      this.routerService.navigate(['/customers']);
+    }
   }
 
   ngOnInit(): void {}